Include a check that the joint angle data is valid.
This check that the joint axis data is valid before using it in the interpolation function. This ensures that the model can load even if the fist time step of the model is not in the joint angle data file. Instead it uses the default data and give an error when the user tries to run the model.