Update shoulder drivers and support ShoulderRhythm
This changes the set of joint drivers for transferring the joint angles from the
over-determinate marker tracking to the inverse dynamic simulation. Instead of
SternoClaviculaElevation, ScapulaThoraxElevation is used, since this makes the
the kinematics of scapula more robust in the inverse dynamic analysis.

The AnyMoCap model now support the shoulder rhythm being enabled.